moved matches and hltvnews to src folder
